Change and Cap Fee Parameters
Introduces a new `RewardParameters` struct to pack fee-related parameters into
a single storage slot, reducing gas costs for reward distribution. It also caps
the fee at 20% to mitigate the risk of excessively high fees being set, addressing
concerns raised during the audit.

+++ test/invariant/UniLst.invariants.t.sol
@@ -13,7 +13,7 @@ contract UniStakerInvariants is Test, UnitTestBase {
   address rewardsNotifier;
   UniLst lst;
   address lstOwner;
-  uint256 initialPayoutAmount = 2500e18;
+  uint80 initialPayoutAmount = 2500e18;
 
   // vars for reducers
   IUniStaker.DepositIdentifier public currentId;
